requirements.txt
appdirs==1.4.4
asgiref==3.3.1
distlib==0.3.1
Django==3.1.5
djangorestframework==3.12.2
filelock==3.0.12
gunicorn==20.0.4
importlib-metadata==3.3.0
PTable==0.9.2
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
virtualenv==20.2.2
virtualenvwrapper-win==1.2.6
whitenoise==5.2.0
zipp==3.4.0
